Create load_cell_reading_histogram.py
Script to test various methods of filtering out outliers.

It appears that rejecting any values that are more than 10g away (in either direction) from both of the past two readings (whether or not those readings were rejected themselves) is an effective way to get rid of outliers and keep only good data.